backend/requirements.txt
alembic==1.13.1
asgiref==3.8.1
bcrypt==4.1.3
blinker==1.8.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.40.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
Django==5.0.6
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-rest-swagger==2.2.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
filelock==3.14.0
future==1.0.0
gevent==24.2.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
idna==3.7
inflection==0.5.1
itsdangerous==2.2.0
itypes==1.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
openapi-codec==1.3.2
packaging==24.0
pillow==10.3.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
PySocks==1.7.1
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
requests-file==2.1.0
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.30
sqlparse==0.5.0
tldextract==5.1.2
typing_extensions==4.12.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
Werkzeug==3.0.3
zope.event==5.0
zope.interface==6.4.post2
